+++ internal/service/service.go
@@ -15,24 +15,13 @@ var (
 	ErrURLAlreadyExist = errors.New("URL already exists")
 )
 
-// URLService определяет контракт для сервиса сокращения URL
-type URLService interface {
-	SetBaseURL(req *http.Request)
-	ShortenURL(originalURL string) (string, error)
-	GetOriginalURL(shortID string) (string, error)
-	GetShortURLIfExists(originalURL string) (string, bool)
-}
-
 // URLServiceImpl - реализация URLService
 type URLServiceImpl struct {
 	storage storage.URLReadWriter
 	baseURL string
 }
 
-// Проверка, что URLServiceImpl реализует URLService
-var _ URLService = (*URLServiceImpl)(nil)
-
-func NewURLService(storage storage.URLReadWriter, baseURL string) URLService {
+func NewURLService(storage storage.URLReadWriter, baseURL string) *URLServiceImpl {
 	return &URLServiceImpl{
 		storage: storage,
 		baseURL: baseURL,
